// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface GoalPageDocumentDataImpactOptionsItem {
  /**
   * value field in *Goal page → Impact options*
   *
   * - **Field Type**: Number
   * - **Placeholder**: Value from 0-5
   * - **API ID Path**: goal_page.impact_options[].value
   * - **Documentation**: https://prismic.io/docs/field#number
   */
  value: prismic.NumberField;

  /**
   * label field in *Goal page → Impact options*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Impact label (eg "Direct positive")
   * - **API ID Path**: goal_page.impact_options[].label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  label: prismic.KeyTextField;
}

/**
 * Content for Goal page documents
 */
interface GoalPageDocumentData {
  /**
   * Impact title field in *Goal page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: How does your solution affect the SDG ##?
   * - **API ID Path**: goal_page.impact_title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  impact_title: prismic.KeyTextField;

  /**
   * Impact options field in *Goal page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: goal_page.impact_options[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  impact_options: prismic.GroupField<
    Simplify<GoalPageDocumentDataImpactOptionsItem>
  >;

  /**
   * Motivation field title field in *Goal page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Motivate why
   * - **API ID Path**: goal_page.motivation_field_title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  motivation_field_title: prismic.KeyTextField;

  /**
   * Motivation field placeholder field in *Goal page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Max 3 000 characters
   * - **API ID Path**: goal_page.motivation_field_placeholder
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  motivation_field_placeholder: prismic.KeyTextField;

  /**
   * Save impact assessment label field in *Goal page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Save assessment
   * - **API ID Path**: goal_page.save_impact_assessment_label
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  save_impact_assessment_label: prismic.KeyTextField;
}

/**
 * Goal page document from Prismic
 *
 * - **API ID**: `goal_page`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type GoalPageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<GoalPageDocumentData>,
    "goal_page",
    Lang
  >;